sentinelex/requirements.txt
cryptography==38.0.4
dnspython==2.2.1
python-whois==0.8.0
bcrypt==4.0.1
argparse==1.4.0
requests==2.28.1



LICENSE 